requirements.txt
gradio
pillow
numpy
websocket-client